agent/requirements.txt
RPi.GPIO==0.7.1
requests==2.31.0
sentry-sdk==1.28.1
